add raw data writer and refactor builder
