+++ waveorder/io/writer.py
@@ -1,9 +1,7 @@
-import numpy as np
 import zarr
 import os
-from numcodecs import Blosc
-from typing import Union
-from zarr import Array, Group
+from waveorder.io.writer_builders import PhysicalZarr, StokesZarr
+
 
 # todo: add checks for inconsistent data shapes
 # todo: change to persistent arrays?
@@ -168,6 +166,8 @@ class WaveorderWriter:
             self.__builder = StokesZarr(self.__builder_name)
         elif self.datatype == 'physical':
             self.__builder = PhysicalZarr(self.__builder_name)
+        elif self.datatype == 'raw':
+            self.__builder = RawZarr(self.__builder_name)
         else:
             raise NotImplementedError("datatype must be one of 'stokes' or 'physical'")
 
@@ -289,548 +289,3 @@ class WaveorderWriter:
 
         self.__builder.set_zarr(self.__current_zarr_dir)
         self.__builder.write(data, t, c, z)
-
-class Builder:
-    """
-    ABC for all builders
-    """
-
-    # create zarr memory store
-    def init_zarr(self,
-                  directory: str,
-                  name: str): pass
-
-    # create subarrays named after the channel
-    def init_array(self,
-                   store: Union[Array, Group],
-                   data_shape: tuple,
-                   chunk_size: tuple,
-                   dtype: str,
-                   pos: int,
-                   prefix: str,
-                   overwrite: bool
-                   ): pass
-
-    # assign group and array attributes
-    def init_meta(self): pass
-
-    # write data to t, c, z slice
-    def write(self,
-              data,
-              t,
-              c,
-              z): pass
-
-
-class PhysicalZarr(Builder):
-    name = None
-
-    def __init__(self, name=None):
-        """
-
-        """
-        self.__pzarr = None
-        self.__compressor = None
-
-        if name:
-            self.name = name
-        else:
-            self.name = 'physical_data'
-
-
-    def write(self, data, t, c, z):
-        """
-        Write data to specified index of initialized zarr array
-
-        :param data: (nd-array), data to be saved. Must be the shape that matches indices (T, C, Z, Y, X)
-        :param t: (list), index or index range of the time dimension
-        :param c: (list), index or index range of the channel dimension
-        :param z: (list), index or index range of the z dimension
-
-        """
-
-        shape = np.shape(data)
-
-        if self.__pzarr.__len__() == 0:
-            raise ValueError('Array not initialized')
-
-        if len(c) == 1 and len(t) == 1 and len(z) == 1:
-
-            if len(shape) > 2:
-                raise ValueError('Index dimensions do not match data dimensions')
-            else:
-                self.__pzarr['array'][t[0], c[0], z[0]] = data
-
-        elif len(c) == 1 and len(t) == 2 and len(z) == 1:
-            self.__pzarr['array'][t[0]:t[1], c[0], z[0]] = data
-
-        elif len(c) == 1 and len(t) == 1 and len(z) == 2:
-            self.__pzarr['array'][t[0], c[0], z[0]:z[1]] = data
-
-        elif len(c) == 1 and len(t) == 2 and len(z) == 2:
-            self.__pzarr['array'][t[0]:t[1], c[0], z[0]:z[1]] = data
-
-        elif len(c) == 2 and len(t) == 2 and len(z) == 2:
-            self.__pzarr['array'][t[0]:t[1], c[0]:c[1], z[0]:z[1]] = data
-
-        elif len(c) == 2 and len(t) == 1 and len(z) == 2:
-            self.__pzarr['array'][t[0], c[0]:c[1], z[0]:z[1]] = data
-
-        elif len(c) == 2 and len(t) == 2 and len(z) == 1:
-            self.__pzarr['array'][t[0]:t[1], c[0]:c[1], z[0]] = data
-
-        elif len(c) == 2 and len(t) == 1 and len(z) == 1:
-            self.__pzarr['array'][t[0], c[0]:c[1], z[0]] = data
-
-        else:
-            raise ValueError('Did not understand data formatting')
-
-    def create_channel_dict(self, chan_name, clim=None):
-
-        if chan_name == 'Retardance':
-            min = 0.0
-            max = 1000.0
-            start = clim[0] if clim else 0.0
-            end = clim[1] if clim else 100.0
-        elif chan_name == 'Orientation':
-            min = 0.0
-            max = np.pi
-            start = clim[0] if clim else 0.0
-            end = clim[1] if clim else np.pi
-
-        elif chan_name == 'Phase3D':
-            min = -10.0
-            max = 10.0
-            start = clim[0] if clim else -0.2
-            end = clim[1] if clim else 0.2
-
-
-        elif chan_name == 'BF':
-            min = 0.0
-            max = 65535.0
-            start = clim[0] if clim else 0.0
-            end = clim[1] if clim else 5.0
-
-        else:
-            min = 0.0
-            max = 65535.0
-            start = clim[0] if clim else 0.0
-            end = clim[1] if clim else 65535.0
-
-        dict_ = {'active': True,
-                'coefficient': 1.0,
-                'color': '808080',
-                'family': 'linear',
-                'inverted': False,
-                'label': chan_name,
-                'window': {'end': end, 'max': max, 'min': min, 'start': start}
-                }
-
-        return dict_
-
-    def set_channel_attributes(self, chan_names: list, clims: list):
-        """
-        A method for creating ome-zarr metadata dictionary.
-        Channel names are defined by the user, everything else
-        is pre-defined.
-
-        Parameters
-        ----------
-        chan_names:     (list) List of channel names in the order of the channel dimensions
-                                i.e. if 3D Phase is C = 0, list '3DPhase' first.
-
-        clims:          (list of tuples) contrast limits to display for every channel
-
-        """
-
-        rdefs = {'defaultT': 0,
-                 'model': 'color',
-                 'projection': 'normal',
-                 'defaultZ': 0}
-
-        multiscale_dict = [{'datasets': [{'path': "array"}],
-                            'version': '0.1'}]
-        dict_list = []
-
-        for i in range(len(chan_names)):
-            if clims == None:
-                dict_list.append(self.create_channel_dict(chan_names[i]))
-            else:
-                #TODO: Check if clims length matches channel length
-                dict_list.append(self.create_channel_dict(chan_names[i], clims[i]))
-
-        full_dict = {'multiscales': multiscale_dict,
-                     'omero': {
-                         'channels': dict_list,
-                         'rdefs': rdefs,
-                         'version': 0.1}
-                     }
-
-        self.__pzarr.attrs.put(full_dict)
-
-    def _zarr_exists(self, path):
-        if os.path.exists(path):
-            print(f'Found existing store at {path}')
-            return True
-        else:
-            # print(f'Creating new store at {path}')
-            return False
-
-    # Placeholder function for future compressor customization
-    def init_compressor(self, compressor_: str):
-        """
-        Zarr supports a variety of compressor libraries:
-            from NumCodecs:
-                Blosc, Zstandard, LZ4, Zlib, BZ2, LZMA
-        Blosc library supports many algorithms:
-                zstd, blosclz, lz4, lz4hc, zlib, snappy
-
-        Parameters
-        ----------
-        compressor_:    (str) one of the compressor libraries from NumCodecs
-
-        Returns
-        -------
-
-        """
-        if compressor_ is "Blosc":
-            self.__compressor = Blosc(cname='zstd', clevel=1, shuffle=Blosc.SHUFFLE)
-        else:
-            raise NotImplementedError("only Blosc library with zstd algorithm is supported")
-
-    # Initialize zero array
-    def init_array(self, store, data_shape, chunk_size, dtype, chan_names, clims, overwrite):
-        """
-
-        Parameters
-        ----------
-        store
-        data_shape
-        chunk_size
-        dtype
-        chan_names
-        clims
-        overwrite
-
-        Returns
-        -------
-
-        """
-        if self.__compressor is None:
-            self.init_compressor('Blosc')
-
-        # Make sure data matches OME zarr structure
-        if len(data_shape) != 5:
-            raise ValueError('Data shape must be (T, C, Z, Y, X)')
-
-        #TODO: GET RID OF THIS?
-        try:
-            self.set_zarr(store[self.name])
-        except:
-            store.create_group(self.name)
-            self.set_zarr(store[self.name])
-
-        self.set_channel_attributes(chan_names, clims)
-        self.__pzarr.zeros('array', shape=data_shape, chunks=chunk_size, dtype=dtype,
-                           compressor=self.__compressor, overwrite=overwrite)
-
-    def init_zarr(self, directory, name=None):
-        """
-        create a zarr.DirectoryStore at the supplied directory
-
-        Parameters
-        ----------
-        directory:  (str) path to directory store
-        name:   (str) name of zarr array, with extension '.zarr'.  Defaults to 'physical_data.zarr'
-
-        Returns
-        -------
-
-        """
-        if name is None:
-            name = 'physical_data.zarr'
-        if not name.endswith('.zarr'):
-            name += '.zarr'
-        path = os.path.join(directory, name)
-
-        # if name is None:
-        #     path = os.path.join(directory, 'physical_data.zarr')
-        # else:
-        #     path = os.path.join(directory, name)
-        #     if not path.endswith('.zarr'):
-        #         path += '.zarr'
-
-        if not os.path.exists(path):
-            store = zarr.open(path)
-            self.set_zarr(store)
-        else:
-            raise FileExistsError(f"zarr already exists at {path}")
-
-    def set_zarr(self, store):
-        """
-        set this object's zarr store
-        Parameters
-        ----------
-        store
-
-        Returns
-        -------
-
-        """
-        self.__pzarr = store
-
-    def get_zarr(self):
-        return self.__pzarr
-
-
-class StokesZarr(Builder):
-    name = None
-
-    def __init__(self, name=None):
-        """
-
-        """
-        self.__szarr = None
-        self.__compressor = None
-
-        if name:
-            self.name = name
-        else:
-            self.name = 'stokes_data'
-
-
-    def write(self, data, t, c, z):
-        """
-        Write data to specified index of initialized zarr array
-
-        :param data: (nd-array), data to be saved. Must be the shape that matches indices (T, C, Z, Y, X)
-        :param t: (list), index or index range of the time dimension
-        :param c: (list), index or index range of the channel dimension
-        :param z: (list), index or index range of the z dimension
-
-        """
-
-        shape = np.shape(data)
-
-        if self.__szarr.__len__() == 0:
-            raise ValueError('Array not initialized')
-
-        if len(c) == 1 and len(t) == 1 and len(z) == 1:
-
-            if len(shape) > 2:
-                raise ValueError('Index dimensions do not match data dimensions')
-            else:
-                self.__szarr['array'][c[0], t[0], z[0]] = data
-
-        elif len(c) == 1 and len(t) == 2 and len(z) == 1:
-            self.__szarr['array'][t[0]:t[1], c[0], z[0]] = data
-
-        elif len(c) == 1 and len(t) == 1 and len(z) == 2:
-            self.__szarr['array'][t[0], c[0], z[0]:z[1]] = data
-
-        elif len(c) == 1 and len(t) == 2 and len(z) == 2:
-            self.__szarr['array'][t[0]:t[1], c[0], z[0]:z[1]] = data
-
-        elif len(c) == 2 and len(t) == 2 and len(z) == 2:
-            self.__szarr['array'][t[0]:t[1], c[0]:c[1], z[0]:z[1]] = data
-
-        elif len(c) == 2 and len(t) == 1 and len(z) == 2:
-            self.__szarr['array'][t[0], c[0]:c[1], z[0]:z[1]] = data
-
-        elif len(c) == 2 and len(t) == 2 and len(z) == 1:
-            self.__szarr['array'][t[0]:t[1], c[0]:c[1], z[0]] = data
-
-        elif len(c) == 2 and len(t) == 1 and len(z) == 1:
-            self.__szarr['array'][t[0], c[0]:c[1], z[0]] = data
-
-        else:
-            raise ValueError('Did not understand data formatting')
-
-    def create_channel_dict(self, chan_name, clim=None):
-
-        if chan_name == 'S0':
-            min = 0.0
-            max = 65535
-            start = clim[0] if clim else 0.0
-            end = clim[1] if clim else 1.0
-
-        elif chan_name == 'S1':
-            min = 10.0
-            max = -10.0
-            start = clim[0] if clim else -0.5
-            end = clim[1] if clim else 0.5
-
-        elif chan_name == 'S2':
-            min = -10.0
-            max = 10.0
-            start = clim[0] if clim else -0.5
-            end = clim[1] if clim else 0.5
-
-        elif chan_name == 'S3':
-            min = -10
-            max = 10
-            start = clim[0] if clim else -1.0
-            end = clim[1] if clim else 1.0
-
-        else:
-            min = 0.0
-            max = 65535.0
-            start = clim[0] if clim else 0.0
-            end = clim[1] if clim else 65535.0
-
-        dict_ = {'active': True,
-                'coefficient': 1.0,
-                'color': '808080',
-                'family': 'linear',
-                'inverted': False,
-                'label': chan_name,
-                'window': {'end': end, 'max': max, 'min': min, 'start': start}
-                }
-
-        return dict_
-
-    def set_channel_attributes(self, chan_names: list, clims: list):
-        """
-        A method for creating ome-zarr metadata dictionary.
-        Channel names are defined by the user, everything else
-        is pre-defined.
-
-        Parameters
-        ----------
-        chan_names:     (list) List of channel names in the order of the channel dimensions
-                                i.e. if 3D Phase is C = 0, list '3DPhase' first.
-
-        clims:          (list of tuples) contrast limits to display for every channel
-
-        """
-
-        rdefs = {'defaultT': 0,
-                 'model': 'color',
-                 'projection': 'normal',
-                 'defaultZ': 0}
-
-        multiscale_dict = [{'datasets': [{'path': "array"}],
-                            'version': '0.1'}]
-        dict_list = []
-
-        for i in range(len(chan_names)):
-            if clims == None:
-                dict_list.append(self.create_channel_dict(chan_names[i]))
-            else:
-                #TODO: Check if clims length matches channel length
-                dict_list.append(self.create_channel_dict(chan_names[i], clims[i]))
-
-        full_dict = {'multiscales': multiscale_dict,
-                     'omero': {
-                         'channels': dict_list,
-                         'rdefs': rdefs,
-                         'version': 0.1}
-                     }
-
-        self.__szarr.attrs.put(full_dict)
-
-    def _zarr_exists(self, path):
-        if os.path.exists(path):
-            print(f'Found existing store at {path}')
-            return True
-        else:
-            # print(f'Creating new store at {path}')
-            return False
-
-    # Placeholder function for future compressor customization
-    def init_compressor(self, compressor_: str):
-        """
-        Zarr supports a variety of compressor libraries:
-            from NumCodecs:
-                Blosc, Zstandard, LZ4, Zlib, BZ2, LZMA
-        Blosc library supports many algorithms:
-                zstd, blosclz, lz4, lz4hc, zlib, snappy
-
-        Parameters
-        ----------
-        compressor_:    (str) one of the compressor libraries from NumCodecs
-
-        Returns
-        -------
-
-        """
-        if compressor_ is "Blosc":
-            self.__compressor = Blosc(cname='zstd', clevel=1, shuffle=Blosc.SHUFFLE)
-        else:
-            raise NotImplementedError("only Blosc library with zstd algorithm is supported")
-
-    # Initialize zero array
-    def init_array(self, store, data_shape, chunk_size, dtype, chan_names, clims, overwrite):
-        """
-
-        Parameters
-        ----------
-        store
-        data_shape
-        chunk_size
-        dtype
-        chan_names
-        clims
-        overwrite
-
-        Returns
-        -------
-
-        """
-        if self.__compressor is None:
-            self.init_compressor('Blosc')
-
-        # Make sure data matches OME zarr structure
-        if len(data_shape) != 5:
-            raise ValueError('Data shape must be (T, C, Z, Y, X)')
-
-        try:
-            self.set_zarr(store[self.name])
-        except:
-            store.create_group(self.name)
-            self.set_zarr(store[self.name])
-
-        self.set_channel_attributes(chan_names, clims)
-        self.__szarr.zeros('array', shape=data_shape, chunks=chunk_size, dtype=dtype,
-                           compressor=self.__compressor, overwrite=overwrite)
-
-    def init_zarr(self, directory, name=None):
-        """
-        create a zarr.DirectoryStore at the supplied directory
-
-        Parameters
-        ----------
-        directory:  (str) path to directory store
-        name:   (str) name of zarr array, with extension '.zarr'.  Defaults to 'physical_data.zarr'
-
-        Returns
-        -------
-
-        """
-        if name is None:
-            name = 'physical_data.zarr'
-        if not name.endswith('.zarr'):
-            name += '.zarr'
-        path = os.path.join(directory, name)
-
-        if not os.path.exists(path):
-            store = zarr.open(path)
-            self.set_zarr(store)
-        else:
-            raise FileExistsError(f"zarr already exists at {path}")
-
-    def set_zarr(self, store):
-        """
-        set this object's zarr store
-        Parameters
-        ----------
-        store
-
-        Returns
-        -------
-
-        """
-        self.__szarr = store
-
-    def get_zarr(self):
-        return self.__szarr
